+++ app/Filament/Resources/Galleries/Schemas/GalleryForm.php
@@ -33,8 +33,9 @@ class GalleryForm
                     ])
                     ->columnSpanFull(),
                 FileUpload::make('image')->label('Foto')
-                    ->helperText('Unggah foto dokumentasi kegiatan (Otomatis dioptimalkan, maks 2MB).')
+                    ->helperText('Unggah foto dokumentasi kegiatan (Otomatis dioptimalkan WebP, maks 2MB).')
                     ->image()
+                    ->acceptedFileTypes(['image/jpeg', 'image/png', 'image/webp'])
                     ->imageResizeTargetWidth(1200)
                     ->maxSize(2048)
                     ->directory('galleries')