+++ app/Filament/Resources/Officials/Schemas/OfficialForm.php
@@ -61,8 +61,9 @@ class OfficialForm
                     ])
                     ->columnSpanFull(),
                 FileUpload::make('photo')->label('Foto')
-                    ->helperText('Unggah pasfoto resmi aparatur (Otomatis dioptimalkan, maks 2MB).')
+                    ->helperText('Unggah pasfoto resmi aparatur (Otomatis dioptimalkan WebP, maks 2MB).')
                     ->image()
+                    ->acceptedFileTypes(['image/jpeg', 'image/png', 'image/webp'])
                     ->imageResizeTargetWidth(600)
                     ->maxSize(2048)
                     ->nullable()